optimizing query, refactoring
